refactor: move opt package into fp
